pyproject.toml
[project]
name = "airflow-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor Romano",email = "victor.romano@traivefinance.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "boto3 (>=1.41.5,<2.0.0)",
    "apache-airflow (>=3.1.3,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "sodapy (>=2.2.0,<3.0.0)",
    "psycopg[binary] (>=3.2.13,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
